float changed into long...

+++ qap/climblarge.h
@@ -4,11 +4,11 @@
 #include "methods.h"
 #include <algorithm>   
 
-float large_first(int ** D, int ** F, vector<int> * sol);
+long large_first(int ** D, int ** F, vector<int> * sol);
 
-float large_best(int** D, int ** F, vector<int> * sol);
+long large_best(int** D, int ** F, vector<int> * sol);
 
-float large_worst(int** D, int ** F, vector<int> * sol);
+long large_worst(int** D, int ** F, vector<int> * sol);
 
 
 #endif
