+++ qap/climbn.h
@@ -4,12 +4,12 @@
 #include "methods.h"
 #include <algorithm>   
 
-float first(int ** D, int ** F, vector<int> * sol);
+long first(int ** D, int ** F, vector<int> * sol);
 
-float best(int** D, int ** F, vector<int> * sol);
+long best(int** D, int ** F, vector<int> * sol);
 
-float worst(int** D, int ** F, vector<int> * sol);
+long worst(int** D, int ** F, vector<int> * sol);
 
-float clone_first(int ** D, int ** F, vector<int> * sol);
+long clone_first(int ** D, int ** F, vector<int> * sol);
 
 #endif
