+++ qap/parser.h
@@ -13,7 +13,7 @@ void parse_qap(int*** D, int*** F, int * size, string filename);
 
 vector<string> tokenize(string toSplit, string token);
 
-void write_res(string fileName, vector<float> iter);
+void write_res(string fileName, vector<long> iter);
 
 void write_res2(string fileName, vector<pair<int,int> > iter) ;
